Consolidate Supervisor API middlewares
