+++ src/lib/api-keys.ts
@@ -1,10 +1,7 @@
 import * as _ from 'lodash';
-import * as express from 'express';
 import * as memoizee from 'memoizee';
 
-import * as config from '../config';
 import * as db from '../db';
-
 import { generateUniqueKey } from './register-device';
 
 export class KeyNotFoundError extends Error {}
@@ -46,7 +43,7 @@ type ScopeTypes = {
 /**
  * The resources which can be protected with scopes.
  */
-interface ScopedResources {
+export interface ScopedResources {
 	apps: number[];
 }
 
@@ -75,19 +72,6 @@ export const isScoped = (
 		scopeChecks[scope.type](resources, (scope as unknown) as any),
 	);
 
-export type AuthorizedRequest = express.Request & {
-	auth: {
-		isScoped: (resources: Partial<ScopedResources>) => boolean;
-		apiKey: string;
-		scopes: Scope[];
-	};
-};
-export type AuthorizedRequestHandler = (
-	req: AuthorizedRequest,
-	res: express.Response,
-	next: express.NextFunction,
-) => void;
-
 // empty until populated in `initialized`
 export let cloudApiKey: string = '';
 
@@ -99,87 +83,8 @@ export const initialized = (async () => {
 	await generateCloudKey();
 })();
 
-/**
- * This middleware will extract an API key used to make a call, and then expand it out to provide
- * access to the scopes it has. The `req` will be updated to include this `auth` data.
- *
- * E.g. `req.auth.scopes: []`
- *
- * @param req
- * @param res
- * @param next
- */
-export const authMiddleware: AuthorizedRequestHandler = async (
-	req,
-	res,
-	next,
-) => {
-	// grab the API key used for the request
-	const apiKey = getApiKeyFromRequest(req) ?? '';
-
-	// store the key in the request, and an empty scopes array to populate after resolving the key scopes
-	req.auth = {
-		apiKey,
-		scopes: [],
-		isScoped: (resources) => isScoped(resources, req.auth.scopes),
-	};
-
-	try {
-		const conf = await config.getMany(['localMode', 'unmanaged']);
-
-		// we only need to check the API key if managed and not in local mode
-		const needsAuth = !conf.unmanaged && !conf.localMode;
-
-		// no need to authenticate, shortcut
-		if (!needsAuth) {
-			// Allow requests that do not need auth to be scoped for all applications
-			req.auth.isScoped = () => true;
-			return next();
-		}
-
-		// if we have a key, find the scopes and add them to the request
-		if (apiKey && apiKey !== '') {
-			await initialized;
-			const scopes = await getScopesForKey(apiKey);
-
-			if (scopes != null) {
-				// keep the scopes for later incase they're desired
-				req.auth.scopes.push(...scopes);
-				return next();
-			}
-		}
-
-		// we do not have a valid key...
-		return res.sendStatus(401);
-	} catch (err) {
-		console.error(err);
-		res.status(503).send(`Unexpected error: ${err}`);
-	}
-};
-
 function isEqualScope(a: Scope, b: Scope): boolean {
 	return _.isEqual(a, b);
-}
-
-function getApiKeyFromRequest(req: express.Request): string | undefined {
-	// Check query for key
-	if (req.query.apikey) {
-		return req.query.apikey;
-	}
-
-	// Get Authorization header to search for key
-	const authHeader = req.get('Authorization');
-
-	// Check header for key
-	if (!authHeader) {
-		return undefined;
-	}
-
-	// Check authHeader with various schemes
-	const match = authHeader.match(/^(?:ApiKey|Bearer) (\w+)$/i);
-
-	// Return key from match or undefined
-	return match?.[1];
 }
 
 export type GenerateKeyOptions = { force: boolean; scopes: Scope[] };
